sol punto 1

+++ src/main/java/it/polito/tdp/yelp/model/Adiacenza.java
@@ -1,15 +1,18 @@
 package it.polito.tdp.yelp.model;
 
 public class Adiacenza {
-
 	
-	String u1;
-	String u2;
-	int peso;
+	 String u1;
+	
+	 String u2;
+	
+	 double peso;
 	
 	
 	
-	public Adiacenza(String u1, String u2, int peso) {
+	
+	
+	public Adiacenza(String u1, String u2, double peso) {
 		super();
 		this.u1 = u1;
 		this.u2 = u2;
@@ -27,12 +30,14 @@ public class Adiacenza {
 	public void setU2(String u2) {
 		this.u2 = u2;
 	}
-	public int getPeso() {
+	public double getPeso() {
 		return peso;
 	}
-	public void setPeso(int peso) {
+	public void setPeso(double peso) {
 		this.peso = peso;
 	}
 	
 	
+	
+
 }
